dynamic_loader support custom device

+++ paddle/phi/backends/custom/custom_device.cc
@@ -1033,39 +1033,44 @@ class CustomDevice : public DeviceInterface {
                       void** blas_handle,
                       phi::stream::stream_t stream) {
     const auto device = &devices_pool[dev_id];
-    CHECK_PTR(pimpl_->init_blas_handle);
-    PADDLE_ENFORCE_CUSTOM_DEVICE_SUCCESS(
-        pimpl_->init_blas_handle(device,
-                                 reinterpret_cast<C_BLASHandle*>(blas_handle),
-                                 reinterpret_cast<C_Stream>(stream)));
+    if (pimpl_->init_blas_handle) {
+      PADDLE_ENFORCE_CUSTOM_DEVICE_SUCCESS(
+          pimpl_->init_blas_handle(device,
+                                   reinterpret_cast<C_BLASHandle*>(blas_handle),
+                                   reinterpret_cast<C_Stream>(stream)));
+    }
   }
 
   void BlasSetMathMode(size_t dev_id, void* blas_handle, int math_mode) {
     const auto device = &devices_pool[dev_id];
-    CHECK_PTR(pimpl_->blas_set_math_mode);
-    PADDLE_ENFORCE_CUSTOM_DEVICE_SUCCESS(pimpl_->blas_set_math_mode(
-        device, reinterpret_cast<C_BLASHandle>(blas_handle), math_mode));
+    if (pimpl_->blas_set_math_mode) {
+      PADDLE_ENFORCE_CUSTOM_DEVICE_SUCCESS(pimpl_->blas_set_math_mode(
+          device, reinterpret_cast<C_BLASHandle>(blas_handle), math_mode));
+    }
   }
 
   void InitBlasLtHandle(size_t dev_id, void** blaslt_handle) {
     const auto device = &devices_pool[dev_id];
-    CHECK_PTR(pimpl_->init_blaslt_handle);
-    PADDLE_ENFORCE_CUSTOM_DEVICE_SUCCESS(pimpl_->init_blaslt_handle(
-        device, reinterpret_cast<C_BLASLtHandle*>(blaslt_handle)));
+    if (pimpl_->init_blaslt_handle) {
+      PADDLE_ENFORCE_CUSTOM_DEVICE_SUCCESS(pimpl_->init_blaslt_handle(
+          device, reinterpret_cast<C_BLASLtHandle*>(blaslt_handle)));
+    }
   }
 
   void DestroyBlasHandle(size_t dev_id, void* blas_handle) {
     const auto device = &devices_pool[dev_id];
-    CHECK_PTR(pimpl_->destroy_blas_handle);
-    PADDLE_ENFORCE_CUSTOM_DEVICE_SUCCESS(pimpl_->destroy_blas_handle(
-        device, reinterpret_cast<C_BLASHandle>(blas_handle)));
+    if (pimpl_->destroy_blas_handle) {
+      PADDLE_ENFORCE_CUSTOM_DEVICE_SUCCESS(pimpl_->destroy_blas_handle(
+          device, reinterpret_cast<C_BLASHandle>(blas_handle)));
+    }
   }
 
   void DestroyBlasLtHandle(size_t dev_id, void* blaslt_handle) {
     const auto device = &devices_pool[dev_id];
-    CHECK_PTR(pimpl_->destroy_blaslt_handle);
-    PADDLE_ENFORCE_CUSTOM_DEVICE_SUCCESS(pimpl_->destroy_blaslt_handle(
-        device, reinterpret_cast<C_BLASLtHandle>(blaslt_handle)));
+    if (pimpl_->destroy_blaslt_handle) {
+      PADDLE_ENFORCE_CUSTOM_DEVICE_SUCCESS(pimpl_->destroy_blaslt_handle(
+          device, reinterpret_cast<C_BLASLtHandle>(blaslt_handle)));
+    }
   }
 
  private:
